Windows,Android: mobile-install fully works
One of the output files in the incremental apk
(stub_application_data.txt) was using CRLF on
Windows, so the mobile-install'ed app was
crashing on startup.

Fix is to open the output file in binary mode so
line endings are not converted to the
host-platform-native one.

See https://github.com/bazelbuild/bazel/issues/3264

Change-Id: Id7d4b5aa4362a21e699517b97dd24858c396eaa7
PiperOrigin-RevId: 164722314