Remove --experimental_dynamic_configs=off.
This is the first in a series of changes stripping out Bazel's
static configuration code.

This change removes the ability to request static configurations but
not the (now orphaned) logic behind them. Because that logic is
all over the code base, it will be removed in layers in followup
changes.

PiperOrigin-RevId: 164769846