Rewrite junitrunner integration tests
Use $(location) to get the unittest.bash script, which allows this test to also
run with a different label prefix. This involves removing the intermediate
testenv.sh script, which doesn't seem to add much value.

We might want to consider a sh_unit_test rule or macro that has an implicit
dependency on the unittest.bash script and doesn't require every test to do
this manually.

PiperOrigin-RevId: 164846477