Rewrite //:combine_distfiles.sh in Python
It can only pack to zip for now (packing to tar is
not trivial and I haven't figured it out yet).

This allows building //:bazel-distfile on Windows.
Previously it was either timing out or taking so
long that it was unbearable (over 10 minutes).
I never waited long enough to see it build.

The new Python version runs under just a few
seconds.

Change-Id: I3264eb7132dd58c581c4216e5bbab035a79d716d
PiperOrigin-RevId: 164954162