Use all resource types in PlaceholderIdFieldInitializerBuilder
All resource types should be supported by PlaceholderIdFieldInitializerBuilder.
Will keeping the ordering from the previous scheme, ensure that all of the
contents of ResourceType.values() are considered. This means that, when adding
new resource types, only updating the ResourceType enum is needed.

A previous attempt at this change tried eliminating the special ordering of
resource types entirely, on the grounds that it shouldn't be needed. However,
strange build failures mean we're keeping it there for now.

RELNOTES: none
PiperOrigin-RevId: 164983075