Send Bazel startup options to server.
Send the startup options tagged with their origin so that the server has correct information about the command line as the client received it.

Removes the unconditional stderr printing of all bazelrc startup options in the bazel client. Instead, the startup options are sent to the server and the same informational printing is gated on the --announce_rc option. This avoids unconditional log spam to stderr early in startup. If the server is unreachable or there are errors parsing startup options, the message is still printed to stderr.

Fixes https://github.com/bazelbuild/bazel/issues/2530.

RELNOTES: --announce_rc now controls whether bazelrc startup options are printed to stderr.
PiperOrigin-RevId: 165211007