Fetch the Java-specific Make variables from the :host_jdk dependency instead of the configuration in genrules.
This is necessary because if one uses a java_runtime rule that has java_home="$(VAR") and VAR is set to an absolute path, BuildConfiguration won't be able to resolve VAR (since it's a Make variable and thus can't affect other Make variables), Blaze won't be able to tell that it's an absolute value and thus will prepend the package name of the java_runtime rule to it, e.g.  resulting in a//foo/bar instead of /foo/bar if the java_runtime rule is in package a.

RELNOTES: None.
PiperOrigin-RevId: 165555251