remote: don't fail build if upload fails
If the upload of local build artifacts fails, the build no longer fails
but instead a warning is printed once. If --verbose_failures is
specified, a detailed warning is printed for every failure.

This helps fixing #2964, however it doesn't fully fix it due to timeouts
and retries slowing the build significantly.

Also, add some other tests related to fallback behavior.

Change-Id: Ief49941f9bc7e0123b5d93456d77428686dd5268
PiperOrigin-RevId: 165938874