Remove tests from //tools/build_defs/docker
The real docker rules are in a separate repo now; docker_build is
maintained here solely for the //src/test/docker tests.

The tests under //tools/build_defs/docker are incompatible with some
desired changes to pkg_tar, and since they now live in a separate repo,
fixing the tests here is not worth the effort, so just remove them
instead.

Change-Id: I642bcdb57382eea1884713ae39f5b55ad5133139
PiperOrigin-RevId: 166031190