Make the srcs test be a Bazel test
./compile.sh srcs has been broken at several time because this
test cannot run on our internal presubmit (more precisely, don't see
most breakage on our internal presubmit). Making this test a Bazel
test should fix that at the same time reducing the scope of ./compile.sh
(with the other move of the determinism test, this make compile.sh
being only the bootstrapping script).

To do so, we fetch the list of sources in the repository through
a skylark repository and use genquery to interrogate the list of
sources in //:srcs filegroup.

Change-Id: If83af28a366cc0c52a7327ce9ed2489c7e88532f
PiperOrigin-RevId: 166064682