Allow CommandLine expansion to throw an exception.
This paves the way for Skylark-side compact command lines that can fail during expansion.

In general, expansion happens in these scenarios:

* Action execution expands the command line to execute it. This is fine since we are well equipped already to handle failing actions.

* In the analysis phase we expand command lines to investigate whether we need a params file. This could be moved to the execution phase, which would have the benefit of getting params files out of the action graph and saving memory.

* Key computation expands the command line. This could be fixed by allowing command lines to compute their own keys (which wouldn't try to expand the command line). This could have the benefit of being more efficient.

* Extra actions expand the command line to construct the extra action proto. This could maybe be deferred to the execution phase (writing the extra action), which would also be more memory efficient.

For failed key computations, we return a singleton "error" key. This means multiple actions that will fail will map to the same key. These actions will necessarily fail if executed, so we should not need to worry about these ending up in the action cache. If we do manage to make the command line compute its own keys then this will become moot in the future.

RELNOTES: None
PiperOrigin-RevId: 166266385