Automated rollback of commit bb0aba3e91b603bf96d4e214a3886d1f47c3d90a.
*** Reason for rollback ***

Publishing duplicate build events for tests.

*** Original change description ***

Request test artifacts to be built in parallel with running the test.

In cases where not all test artifacts are needed to run the test, this allows for greater parallelism in the build.

We need to inject the request for the test to be run into the TargetCompletion function, so that it can properly process any errors. Thanks to nharmata@ for suggesting this.

MEMORY: One additional Skyframe node and 2 edges for each test to be run (sharded tests still only count as one). So with 500 test target...

***

PiperOrigin-RevId: 166307568