Refactor collection of action cache statistics.
This change gets rid of the old CachesSavedEvent object (note, plural) which
used to capture details about various caches.  As this was now only used for
a single cache, rename it to ActionCacheStatistics so that the data it
contains is more cohesive.

The reason for this change is to make room for the addition of other metrics
for the action cache (like hits/misses), which will come later.  As of now,
this change is intended to be a no-op.

While doing this, use AutoValue to avoid mutability of the generated objects
and use better types for the contained data (e.g. Duration for timings).

RELNOTES: None.
PiperOrigin-RevId: 166742117