Fix NPE in WorkerFilesHash.
We must check that a given artifact is a file, before calling getDigest, according to the interface contract specified in actions/Metadata.java.

Add a regression test to bazel_worker_test.sh, too. It's enough to simply add a directory entry to the worker's data to trigger the original bug.

PiperOrigin-RevId: 166856512