Save density information to manifests when filtering resources in analysis
Previously, when filtering in analysis, density information was not passed to
execution to avoid wasting time on re-filtering. However, that density
information was also used to add information to the manifest. Add an additional
flag to save density information to the manifest without triggering resource
filtering in execution, and use it when filtering in analysis.

RELNOTES: None
PiperOrigin-RevId: 166891827