Fix .dwo Artifact construction with ThinLTO+Fission for nativedeps
The code in scheduleLtoBackendAction was attempting to construct the Artifact for a .dwo file when Fission is enabled by getting the related artifact via the ruleContext. This doesn't work for nativedeps files which have a shared _nativedeps/-relative library path, instead of being under the package directory.

Instead, compute the .dwo path the same way we do the ThinLTO imports and index file outputs, using the linkArtifactFactory.

PiperOrigin-RevId: 167002233