Adjust Remote TreeNodeRepository scope to build
Prevent massive numbers of redundant filesystem interactions and digest
calculations by rescoping RemoteSpawnStrategy from per-exec to
per-build.

synchronize treeNodeDigestCache access

TreeNodeRepository has been observed to return null for getMerkleDigest,
which indicated a race condition between member maps.  By synchronizing
all access to the treeNodeDigestCache member, it will be impossible for
uncoordinated updates of the member maps which determine cache presence
and consistency to occur without synchronization of the accessors.

In this case, fileNodeCache could have been updated with a key that had
not yet been inserted into treeNodeDigestCache, with a yield resulting
in a cache 'hit' with inconsistent data in the member caches, and a
resulting fetch of the still unavailable correlative data in
treeNodeDigestCache.