Android,Windows: wrap SingleJar in a java_binary
On Unix-like platforms, we can execute deploy jars
just like normal binaries, but on Windows we
cannot because ".jar" is not an executable
extension. Therefore we cannot run
SingleJar_deploy.jar in a genrule.cmd.

In this change I do the following changes to:
//tools/android/BUILD.tools:
- add a java_import rule that wraps the SingleJar
  deploy jar
- add a java_binary that runtime-depends on the
  jar
- fix the ":gen_java_lang_extras_jar" genrule to
  run the java_binary instead of the deploy jar

See https://github.com/bazelbuild/bazel/issues/3659

Change-Id: Ie035e721deb696bbd4130b3612695a10dd9c2b4d
PiperOrigin-RevId: 167281195