Remove the product name from ConfiguredRuleClassProvider.
The RuleClassProvider includes a copy of the product name, parameterized
for both Blaze and Bazel.  Apparently, this is exclusively there so that
the standalone docgen binary can "magically" guess the product name.

This is strange and adds additional complexity to the Bazel core codebase
for no strong reason.  Instead, just add a new flag to docgen that takes
the product name and pass it in explicitly.

RELNOTES: None.
PiperOrigin-RevId: 167724033