Replace ad hoc wrapper classes with Equivalence.Wrapper
The tradeoff is that it eliminates a class but makes type signatures slightly more complicated. I think it's a net win because we avoid implementing equals/hashCode.

I figured helper methods for wrapping/unwrapping were better than relying on the caller to know what Equivalence to use ("identity").

RELNOTES: None
PiperOrigin-RevId: 167750079