Don't attempt LTO indexing of cc_binary if there are no bitcode inputs to link.
The problem is that the linker will not produce the expected .params file output
if there are no inputs, and there is no need for this step in that case.
This required adding a method to check if there are any bitcode inputs to
the link either directly or indirectly via a library. This is the same walk done
by createLtoArtifacts once we decide to do an LTO indexing step.

RELNOTES: NONE
PiperOrigin-RevId: 168281516