remote: Add new retrier with support for circuit breaking
Add a generic retrier implementation (Retrier2) that can be
configured by plugging in a backoff strategy, a function to
decide on retriable errors and a circuit breaker. A concrete
implementation is added via RemoteRetrier that mostly is a
copy of the code of the existing Retrier.

Retrier2 adds support for circuit breaking [1]. It allows the
retrier to reject execution when failure rates are high. The
remote execution code will use this to gently switch between
local and remote execution/caching if the latter experiences
lots of failures.

Retrier2 is also useful when not used with gRPC. We need
retriers for the HTTP caching interface too.

All the code added in this CL is unused, to keep reviews
managable. In a follow up CL, I will switch the code to use
the new Retrier and delete the old retrier.

[1] https://martinfowler.com/bliki/CircuitBreaker.html

PiperOrigin-RevId: 168355597