Merging a RunfilesSupport object into Runfiles.Builder no longer adds the runfiles middleman to the artifacts.
Runfiles.Builder.merge(RunfilesSupport) would add the runfiles
middleman of the RunfilesSupport object being merged into the
artifacts of the nascent Runfiles object. This effectively made the
construction of one runfiles tree depend on that of another.

Instead of putting the runfiles middleman into the runfiles artifacts,
we can put it into the FilesToRunProvider, which should have the same
effect. To facilitate this, RuleConfiguredTargetBuilder gains a
addFilesToRun method. Note that not every caller of
merge(RunfilesSupport) has to be modified because some were already
adding the runfiles middleman to their FilesProvider (e.g.,
AndroidHostServiceFixture), which is a subset of FilesToRunProvider.

This is split out from https://bazel-review.googlesource.com/c/14010

Change-Id: Icfca5262a32619e1e351477f3ff641350b6798dc
PiperOrigin-RevId: 168358030