bes: cancel stream on bazel shutdown
Introduce a new method BuildEventTransport#closeNow() which forcefully
closes the BES upload even if there are still events to upload. This
is executed if the bazel server is being shutdown.

The implementation in the BuildEventServiceTransport works by shutting
down the ExecutorService handling the upload, which in turn will
interrupt the upload thread, which in turn will cancel the streaming RPC.

The BuildEventServiceTransport will wait up to 100ms for that
cancellation to happen.

Also fix a bug where the BES transport would not wait for ACKS.

PiperOrigin-RevId: 168359947