Automated rollback of commit 72de1fe6bafa28fddcbc8cdc1ccbb59556996bfb.
*** Reason for rollback ***

Roll-forward with fix

*** Original change description ***

Automated rollback of commit f5b8281f1f7599c67476663887db909a4206710e.

*** Reason for rollback ***

Read the wrong log, it is not timing out, actual failure.

*** Original change description ***

Open-source src/test/skylark/...

This was means to be open-sourced from the beginning but an error in our
set-up hide it from the open-source tree.

PiperOrigin-RevId: 168526758