Automated rollback of commit ca216f5eb4b1f1892219db0f578495bbbdbaa875.
*** Reason for rollback ***

Breaks Bazel building itself on FreeBSD, also #3739.

*** Original change description ***

Introduce empty "toolchain_category" rule for labels that will be used as
categories of toolchains for the purpose of toolchain selection.

Up to now, we've used the native toolchain_type rule for this purpose.  That rule depends on a number of configuration fragments that supply build variables - we don't want toolchains to need to depend on those fragments as well.  E.g. toolchain_type depends on JvmConfiguration, but we would like toolchains to work with --experimental_disable_jvm.

PiperOrigin-RevId: 168810566