Java Launcher: Create classpath jar with a random name
When the classpath is too long, the launcher creates a jar file to pass
the classpath value. This change makes the jar file's name random.

It fixed the bug that when running multiple instances of the same java binary,
they all try to create the same classpath jar file. For example, when
running java_test with shard_count > 1.

Also, we delete the classpath jar file after the program finishes, so
that we don't leave any garbages behind.

Change-Id: I67926b3ef76dcb1806797e977ecaa7c6763c5cf2
PiperOrigin-RevId: 169087032