Make desugar tests assume android-24.
Previously, they assumed android-25 but Bazel CI does not have android-25
installed. In practice, these tests do not care which version is used so we
should use //tools/android:defaults_jar which selects the jar based on the
configuration. However, that target does not produce the jar as a runfile, it
is only useful currently for neverlink compilation.

RELNOTES: None
PiperOrigin-RevId: 169104539